requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
blinker==1.8.2
cachelib==0.9.0
certifi==2024.6.2
click==8.1.7
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.111.0
fastapi-cli==0.0.4
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
libcst==0.4.10
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
motor==3.4.0
mypy-extensions==1.0.0
nltk==3.8.1
numpy==2.0.0
orjson==3.10.5
packaging==24.1
pandas==2.2.2
pydantic==2.7.4
pydantic_core==2.18.4
Pygments==2.18.0
pymongo==4.7.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
remove-print-statements==0.5.2
rich==13.7.1
scikit-learn==1.5.0
scipy==1.13.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
threadpoolctl==3.5.0
tqdm==4.66.4
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
Werkzeug==3.0.3
